validOceanMask updated based on availOceanMask
validOceanMask is now updated based on availOceanMask at the end of the
vertical extrap routine, instead of bed topography. This is the original
version of the code. availOceanMask is already defined as being above
bed topography so this shouldn't make a difference in performance, but
will be more consistent with the original implementation.