Insert addition vert extrap for 200-500
For the 200-500 vertical TF parameterization, a vertical extrapolate is
now performed prior to the horizontal extrapolation. This ensure proiles
extend to the deepest vertical layer before horizontal extrapolation,
making sure that horizontally extrapolated water properties are as local
as possible and not from far field cells when only a few valid cells at
a particular vertical level exist.