Add regional stats output for MALI
This commit adds regional stats analysis member history files for MALI.
They follow the convention of the recently modified global stats files
to have a file interval of 1000 years.

Regional stats requires a region mask input file to be useful, which is
not available (or necessarily appropriate) for all currently supported MALI grids.
Logic is added to only input a region mask for grids for which a region
mask exists.  For older grids that don’t have a region mask file, the
regional stats will be calculated for a single region covering the
entire domain, making this redundant with the global stats, but not
harmful.  Region mask files have been added to the input server for the
grids for which they exist, following the naming convention used here.