Add DROF options to zero rofl or rofi around AIS or GIS
This commit adds new DROF namelist options to remove either rofl or rofi
from around either the Antarctic Ice Sheet or the Greenland Ice Sheet.
This is to be used when those runoff fluxes are represented elsewhere -
either by data icebergs or an active GLC component.

Logic is added in DROF's buildnml to activate the AIS options following
the same rules currently for removing runoff currently implemented in MPAS-Ocean.
The Greenland options are not currently used.