Use a connected ocean mask to avoid subglacial lakes in ismip ISMF calc
This moves the calculation of floating cells connected to the open ocean
to an earlier part of the iceshelf_melt module and saves it into a new
Registry variable.  That new variable is then used within the ismip6
ice-shelf basal melt routine to prevent floating ice over subglacial
lakes from being included in melt calculations.

This is needed for TF coupling to E3SM where those locations have
no_data values that corrupt TFmean.  But it also corrects a minor bug in
standalone ismip-type configurations where the TF in these floating lake
regions was contributing to TFmean (but, note, no melt was being applied
in those lakes because they were masked out when melt was applied).

Note that there is a similar Registry variable used in the extrapolation
code called connectedMarineMask.  It is calculated the same way and
these two fields are redundant.  I chose to keep them separate to avoid
complexity in these two modules using the same field and needing to
assume it was up to date.  I gave it a slightly different name even
though they belong to different pools because i/o requires all Registry
variables have unique names even if they belong to different pools.