Delay modification of ocean state for some sequences
For coupler sequencing options CESM1_MOD, CESM1_MOD_TIGHT,
allow the ocean state to be modified early just so that the o2i
can happen.  Then restore it and really modify it later.  This is
needed to correctly implement those sequencing options the way
driver-mct does it.  The single moab mesh instance per component-in-coupler
was throwing off the timing of the ocean update from what it should be.
This removes diffs wth driver-mct for the C-case.