Pass accumulatedFrazilIceMass between export routines.
Save accumulatedFrazilIceMass in export_mct before it is operated on and
pass it to export_moab using a module variable.  This is needed for driver-moab
to match answers with MCT in the HAVE_MOAB case when both are called. This will be removed
when the export routines are formally seperated.

Also move the MOABDEBUG write of MOAB's ocn_export to the run method so it can also
include diffs with MCT.