Update OCN2GLC_TF_SMAPNAME mapping files for Greenland
Temporary OCN2GLC_TF_SMAPNAME mapping files had been specified for the
Greenland meshes due to known issues with the ocean extrapolation code
in MALI.  Recent fixes to the extrapolation scheme allows the correct
OCN2GLC_TF_SMAPNAME mapping files to be switched back to for the
mpas.gis4to40km mesh.

For the mpas.gis1to10kmR2, there was a different issue requiring
a nonideal OCN2GLC_TF_SMAPNAME mapping file - the gutter on mpas.gis1to10kmR2
is too small to provide sufficient overlap with the MPAS-Ocean mesh.  A
temporary fix had been implemented to use a mapping file with
extrapolation to translate the ocean data over to the MALI mesh.
However, recent testing showed that this fix does not work as expected,
because it also translates ocean depths beneath the bathymetry but there
is not a way to flag them as invalid, so MALI ends up using a bunch of
no data (1e36) values for thermal forcing.  This commit introduces a new
mapping file that excludes any ocean cells where the bathymetry is
shallower than the deepest depth that MALI needs around Greenland to
avoid this issue.  However, this means that MALI is unable to receive
any ocean information from most of the continental shelf, which
obviously is undesirable.  This remains the least bad option for the TF
coupling on the mpas.gis1to10kmR2 grid until a mpas.gis1to10kmR2 grid
with a larger gutter is created.