Perform file existence check on master proc in ice_pio_init()
In ice_pio_init(), the inquire() call is currently made on all
procs, but the results may not be consistent. Some procs may
return .true. while others return .false., which can lead to
mismatches in the clobber flag passed to SCORPIO.

This update modifies ice_pio_init() to perform the inquire check
only on the master proc and then broadcast the result to all other
procs.