Remove berg package from berg variables
With Fixx_rofi now being coupled to ice, the berg variables will be used
all the time when running in E3SM, as opposed to just when data icebergs
are active.

Note that this is unlikely to be the most desirable solution; the berg
variables would ideally be disabled for standalone MPAS-Seaice, but with
this change they will always be allocated.  I could not find an obvious
way to make the package conditional on MPAS-Seaice being run within
E3SM, which I think would be the preferred criterion.  Also, if the
solution of having the berg variables be always allocated is acceptable,
then there is no longer a need for the berg package and it could be
fully removed.  So either way, this commit will need to be revisited.