Enable coupler var iceberg_prognostic from seaice driver
This coupler flag is what tells the coupler to send rofi fluxes from ROF
and GLC to ICE.  This will allow seaice to act as a pass-through for
rofi fluxes to the ocean.  When an iceberg model is eventually added to
MPAS-Seaice, it can redistribute icebergs before they are passed on to
the ocean.