Fix field list in prep_ice_calc_g2x_ix
Fixx_rofi had been specified in this list when Fixx_rofi coupling  was
partially implemented (but not tested) years ago.  However, that is the
wrong variable at this point in the coupling chain - Figg_rofi is the
field that needs to be remapped at this stage.

Note that usually a previously defined list of variables is used at
this stage, as opposed to hardcoding a variable name/list, but I am
modifying the existing implementation and not changing that here.