Move DIB spreading of coupler Fixx_rofi flux ice_import_mct plus fixups
This commit does 2 things:
* Most obviously, I moved the code that uses the global sum of the
coupler iceberg flux to ice_comp_mct.  I did this because the spreading
was not occurring and I thought there might be a sequencing problem with
it being called from mpas_seaice_forcing.F.  By putting the scalign
directly after the Fixx_rofi field is imported to bergFreshwaterFlux, it
ensures there is not a lag in the scaling being applied.
* Adds areaIntegAnnMeanDataIcebergFreshwaterFlux to the input stream
reading DIB information.  Without this, the
areaIntegAnnMeanDataIcebergFreshwaterFlux field will default to 0, which
could well be the explanation for why the previous commit didn't work right.
So it is possible the scaling could move back to mpas_seaice_forcing.F,
but I feel it is more appropriate to have it occur immediately after
importing the field.  However, this decision should be informed by input
from other developers.

This commit has some to-do items:
* Currently the global sum of bergFreshwaterFlux is used for DIB
scaling.  This ignores the fact that bergFreshwaterFlux could exist for
both ice sheets, as well as the fact that bergFreshwaterFlux is assigned
value from Fixx_rofi, which as designed will include rofi from both GLC
and ROF components (but ROF is not currently hooked up).  It is unclear
if we want the scaling/spreading to use rofi from ROF.  Handling the
other issue - treating AIS and GIS bergFreshwaterFlux separately - is
likely to require a mask or latitude assumption combined with using
multiple DIB files.  Note that if we adopt the Schmidt and Mankoff
DIB-by-region mappings, we could use the region masks to handle the
separation between ice sheets - by appending the regions from the ice
sheets, nothing about the implementation would need to know that a
specific region applies to one ice sheet or the other.  So it might be
worth adding that to this PR and making region masks a requirement.
* indentation needs cleanup