Add field sublist tracking to output
This commit adds to the scorpio output yaml parameter list parsing a
check if a field requested for output has an accompanying parameter
sublist.  If it does it tracks that and in the case of diagnostics
relies on the sublist to specify the diagnostic settings, rather than,
for example, a regex check.

This should make it possible to specify complicated diagnostics
in a more human readable format.

[BFB]