Add a unit test to compare the regex vs. sublist approach for diags.
This commit adds output fields to the regular homme+physics standalone
test that use the regex syntax to specify complicated diagnostics.

A separate stream is defined which has the same set of output fields,
but the complicated diagnostics are defined using the sublist approach.

These are then compared via CPRNC to make sure we get the same results.