Add helper fields to output interface to handle transposed output.
This commit addresses a reviewer suggestion that we can add "helper"
fields to the output interface for cases like transposed output.  We
can then initialize those helper fields, which are defined by field
layout, once and resue the allocated memory when we transpose and
write output.

This commit also updates the checks for the field transpose function
to ignore cases where either of the fields in the transpose operation
have the units set to "invalid"