Fix order of `MaxCellsOnEdge`/`TWO` dimension read
We want to try the new name first, then the old one.  As it stands,
we get an error message like:
```
[error] [IO.cpp:343] PIO error while reading dimension TWO
```
even if `MaxCellsOnEdge` is then successfully read.  It isn't
great that we're getting an error message at all but at least
with this fix the error will only be logged if the old dimension
name is being used.