Prevent uphill outflow at terrestrial margins
This commit fixes a bug that was introduced when downhill terrestrial
outflow was modified to ignore to the downhil bedtopo slope.  This
commit ensures that there is no outflow if the bedtopo external to the
grounded ice is a higher elevation than the bedtopo under the ice.  This
was previously occurring.