Disable iceThicknessHydro option by default
In AIS simulations, having iceThicknessHydro enabled was causing massive
lake formation and lack of channelization.  On further investigation, I
found there are thickened artifacts of increased thickness adjacent to
grounding lines.  This was in effect creating berms in the
hydropotential, making it harder for water to reach the ocean.  The
iceThicknessHydro calculations should be improved to eliminate this
issue, or, if that's not possible, removed from the code.  But for now I
am disabling it by default.

Note that I created an MPAS-Tools script to
preprocess ice thickness to eliminate depressions in the surface
elevation, and this greatly improves model stability around subglacial
lakes.  That preprocessing step should be used instead of
iceThicknessHydro.  It is unclear if that will be sufficient in the case
of evolving ice thickness.  If not, the surface-depression-filling
algorithm it employs should be ported to MALI and replace the existing
iceThicknessHydro algorithm.