EAMxx: change behavior of print_global_state_hash
The routine now only requires info on whether it is called
before or after the run_impl call. If pre-run, hashes
input+internal fields, otherwise output+internal. In the
pre-run call, we only hash internal fields that are part
of the restart group, as the others are assumed to be
diagnostic internal fields, and hence they are allowed to
have different content before the run_impl call