Remove support for 20km Greenland mesh and related configurations.
This commit removes basic support for model configurations that use the
20km, uniform resolution Greenland mesh, which is being replaced with
the var. res. 4to40km Greenland mesh. This commit does NOT
remove/replace support that has already been replaced as part of PR
7892, which updates testing support (and associated PE layouts) to use
the 4to40km init. cond. file instead of the 20km file. The assumption is
that these commits will be integrated sequentially.